# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Claw API",
    "description": ""
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "tags": [
    {
      "name": "Product"
    },
    {
      "name": "User"
    }
  ],
  "paths": {
    "/product/": {
      "get": {
        "tags": [
          "Product"
        ],
        "description": "不傳任何參數，將會回傳全部資料",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "產品名稱",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "產品分類",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "tags": [
          "Product"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/product"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/product"
              }
            }
          }
        }
      }
    },
    "/product/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Product"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "tags": [
          "Product"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/productUpdate"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/productUpdate"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "product": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "可樂果"
          },
          "points": {
            "type": "number",
            "example": 100
          },
          "category": {
            "type": "string",
            "example": "食品"
          }
        },
        "required": [
          "name",
          "points",
          "category"
        ],
        "xml": {
          "name": "product"
        }
      },
      "productQuery": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "可樂果"
          },
          "category": {
            "type": "string",
            "example": "食品"
          }
        },
        "required": [
          "name"
        ],
        "xml": {
          "name": "productQuery"
        }
      },
      "productUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "乖乖"
          },
          "category": {
            "type": "string",
            "example": "零食"
          }
        },
        "xml": {
          "name": "productUpdate"
        }
      }
    }
  },
  "externalDocs": {
    "description": "",
    "url": "http://localhost:3000/api-docs"
  }
}